refactor: separate parser and renderer
